requirements.txt
azure-common==1.1.26
azure-core==1.27.0
azure-functions==1.6.0
azure-servicebus==7.0.1
azure-storage-blob==12.16.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
cryptography==41.0.1
idna==2.10
isodate==0.6.1
msrest==0.6.21
numpy==1.22.2
oauthlib==3.1.0
pandas==1.4.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
requests-oauthlib==1.3.0
salesforce-api==0.1.40
six==1.15.0
typing_extensions==4.6.3
tzdata==2023.3
uamqp==1.5.3
url-normalize==1.4.3
urllib3==1.26.6
xmltodict==0.12.0
